requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
Pillow==10.3.0